minor updates particularly to the sigmoid function

+++ Project1/Network.h
@@ -12,11 +12,11 @@ private:
 	std::vector<Matrix> Biases;
 	std::vector<std::vector<Matrix>> Cost;
 public:
-	Network(std::vector<std::pair<int, int>> arrangement, int range);
+	Network(std::vector<std::pair<int, int>> arrangement, float range);
 	Matrix evaluate();
 	void CalcCostDerivative(Matrix DesiredOutput, int stage);
 	float CalcCost(Matrix DesiredOutput);
 	std::pair<std::pair<int, int>, std::pair<int, int>> Getaw(std::pair<int, int> el1, std::pair<int, int> el2);
 	std::vector<std::pair<int, int>> Bakval(bool A, int i, int j,int extntr,int extntc);
-	void backprop(int cost);
+	void backprop(float cost);
 };
